// miniprogram/api/generated/models/TopicArticleItemDto.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TopicArticleItemDto
 */
export interface TopicArticleItemDto {
    /**
     * 
     * @type {number}
     * @memberof TopicArticleItemDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof TopicArticleItemDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof TopicArticleItemDto
     */
    authorNickname?: string;
    /**
     * 
     * @type {number}
     * @memberof TopicArticleItemDto
     */
    viewCount?: number;
    /**
     * 
     * @type {number}
     * @memberof TopicArticleItemDto
     */
    likeCount?: number;
    /**
     * 
     * @type {number}
     * @memberof TopicArticleItemDto
     */
    rank?: number;
    /**
     * 
     * @type {string}
     * @memberof TopicArticleItemDto
     */
    publishedAt?: string;
    /**
     * 
     * @type {number}
     * @memberof TopicArticleItemDto
     */
    hotScore?: number;
}

/**
 * Check if a given object implements the TopicArticleItemDto interface.
 */
export function instanceOfTopicArticleItemDto(value: object): value is TopicArticleItemDto {
    return true;
}

export function TopicArticleItemDtoFromJSON(json: any): TopicArticleItemDto {
    return TopicArticleItemDtoFromJSONTyped(json, false);
}

export function TopicArticleItemDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TopicArticleItemDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'authorNickname': json['author_nickname'] == null ? undefined : json['author_nickname'],
        'viewCount': json['view_count'] == null ? undefined : json['view_count'],
        'likeCount': json['like_count'] == null ? undefined : json['like_count'],
        'rank': json['rank'] == null ? undefined : json['rank'],
        'publishedAt': json['published_at'] == null ? undefined : json['published_at'],
        'hotScore': json['hot_score'] == null ? undefined : json['hot_score'],
    };
}

export function TopicArticleItemDtoToJSON(json: any): TopicArticleItemDto {
    return TopicArticleItemDtoToJSONTyped(json, false);
}

export function TopicArticleItemDtoToJSONTyped(value?: TopicArticleItemDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'author_nickname': value['authorNickname'],
        'view_count': value['viewCount'],
        'like_count': value['likeCount'],
        'rank': value['rank'],
        'published_at': value['publishedAt'],
        'hot_score': value['hotScore'],
    };
}